+++ app/src/main/java/uk/colessoft/android/hilllist/model/TinyHill.kt
@@ -1,7 +1,7 @@
 package uk.colessoft.android.hilllist.model
 
 class TinyHill {
-    var _id: Int = 0
+    var _id: Long = 0
     var hillname: String? = null
     var latitude: Double? = null
     var longitude: Double? = null
